Update README.md
This Python script defines a function to generate random passwords of variable length. The user is prompted to input the desired password length, and the script utilizes the `random` and `string` modules to create a password consisting of letters (both uppercase and lowercase), digits, and punctuation. The generated password is then printed to the console.

Usage:
- Run the script.
- Input the desired length of the password when prompted.

Example:
Enter the length of the password: 12
Generated password: 1a$BcD*3Efg